+++ src/main/java/core/strategy/OscillatorStrategyImpl.java
@@ -16,7 +16,7 @@ public class OscillatorStrategyImpl implements IOscillatorStrategy {
             return false;
         }
 
-        int val1=l.get(0).getValue(),val2=l.get(1).getValue();
+        double val1 = l.get(0).getValue(), val2 = l.get(1).getValue();
         int marginOfError=10;
         for(int i = 2;i<l.size();i++){
             if(l.get(i).getValue()<=val1+marginOfError && l.get(i).getValue()>=val1-marginOfError){
